New important OS support messages
Because the PDCurses wrapper we were using no longer exists and we
didn't have a backup, Windows support has been dropped for version
0.022.  A message has been added in ``src/sys.d'' warning the user about
this issue.

Additionally, the "unsupported OS" message has been changed to clarify
that the code will most likely "just work" even if it is not on a system
in the active support list.

Still don't use Apple products, though.