Updated the to-do list
Since we don't have a PDCurses wrapper anymore, SDL support has been
added to the to-do list and to-do items having to do with the game's
display have been re-prioritized to the top.

The to-do item having to do with a string formatting wrapper has been
removed because the transition to D has made that irrelevant.

You know, when I opened my Dropbox account I wasn't expecting most of
the "you added a file" notifications to contain git objects, but here we
are!