Add a version control for curses import
Because ncurses and PDCurses are only available on certain systems, a
version control check has been added in ``src/global.d'' to determine
which version should be used.

Because there isn't a PDCurses wrapper, the Windows version displays a
warning notifying the user.  If a non-POSIX, non-Windows system is
detected (I'm not sure what systems those would be but you never know)
a separate warning is displayed.