De-uglified the libtcod license notice
The license notice for the Doryen library (libtcod) has been reformatted
for better readability.  There are more than just cosmetic reasons for
doing this: Originally I had simply copied the comment block from the
field-of-vision code I used, based on a misinterpretation of the wording
of the license, but this included a lot of asterisks which in my opinion
rendered the asterisks used for their bullet point list too confusing.

It's better form to do this, because it shows that I'm actually using
and working with the libtcod source code, and that I'm grateful for
being able to do so, rather than lazily slapping it into my own code
like some kind of barbarian hippie who thinks they can just take
everything they want from someone else.  There are way too many people
like that preying on the open-source community.

Hmm, somewhere along the way this transitioned away from being a commit
message and more into a rant about how inconsiderate some people are.
Hopefully it was funny.