Update to the documentation for version 0.022
The changelog and history documents have been updated.

The readme has been modified to update the style guide, update the
compilation instructions, and notify users of the lack of Windows
support for version 0.022.

The readme has also been reorganized so that people will hopefully get
information in the order they need it: First the e-mail address for the
maintainer, then the introduction, then the licensing notice, then what
systems are supported, the compiling instructions, style guide, and
then the outro.  The general theme for this change was to put the
information that most users would need at the top and then, the further
down one goes, the more in-depth they would need to be with the program
and its source code in order to need it.