Changes to the license agreement
The clause requiring users to submit their modifications to the
copyright holder and/or maintainer upon request has been given a time
limit in the interests of fairness.  A requirement that derivative
software change the name has been added.  There have been some changes
to the wording of the "Letter from the Author."