Add a warning for PDCurses
Due to repeated and ultimately failed attempts to get PDCurses working
in dmd, a warning message has been added to global.d which warns the
user that PDCurses is not being actively supported.  This message
instructs the user to leave an Issue on the official github page if they
figure out a workaround.

Additionally, a message has been added which alerts the user of the
version number being compiled if it is a release version.