SDL setup
The initial setup for the SDL version of the game is complete.  Three
new configuration options are added:

  "default" depends on SDL
  "ncurses-nosdl" depends on ncurses
  "pdcurses-nosdl" depends on pdcurses

Two existing configuration options are modified:

  "ncurses" depends on ncurses AND SDL
  "pdcurses" depends on pdcursos AND SDL

Some changes are made to the source code to allow certain files to
compile with or without curses.

``global.d'' is modified to determine what display options are available
to the Spelunk! program.  `SPELUNK_CURSES' evaluates to true if
pdcurses or ncurses are available.  `GFX_NONE' evaluates to true if SDL
is NOT available.

`main' has been changed to accept command line arguments to enable or
disable SDL, and an enum and a static variable have been defined for
this purpose.  std.getopt is being used to take in command line
arguments.  Depending on the configuration, `SDL_Mode' may have no
effect.

The game still doesn't compile yet because display.d needs to be
modified to accept both curses and SDL, and we need to initialize the
SDL window.