Partial interfacing of drawing functions
``display.d'' has been repurposed so that it creates an `interface' for
a `class' which stores drawing functions and then imports new files
``displayc.d'' and ``displayt.d'' (and later ``displayg.d'' or something
for the graphical version) depending on configuration.

Drawing functions called in the `main' loop in ``spelunk.d'' are changed
to use the new class (actually just CursesDisplay, since SDL isn't
implemented yet) and a few other minor changes have been made here and
there, mostly to optimize function calls.

The interfacing isn't quite done yet, so this won't compile.  Here's the
to-do list for now:

  - Add functions to `Display' which govern displaying the message
    buffer
  - The SDL terminal interface