Included development libraries and Proggy
Development libraries for the versions of SDL2 and SDL_ttf that we're
using for development are inculded with the source code in a ``lib''
directory.  Additionally, development libraries for FreeType and zlib,
which SDL_ttf depend on, are included.

Additionally, the ProggySquareSZ font is included in the ``assets''
folder.

``3rdparty.txt'' has been modified to give attribution for these
dependencies, and license notices have been included for all.

A text file ``lib/unix/sdl.txt'' has been added alerting users where to
go to find Linux and FreeBSD versions of SDL and SDL_ttf.