Rooms now connect randomly
To add a little spice to the maps, rooms are now shuffled before being
connected so that they are connected in a random order rather than in a
straight line.  This makes the maps more interesting, as it is less
obvious in which direction the player is expected to go, and corridors
are more likely to intersect rooms, allowing rooms to have multiple
exits.