Use HILITE_PLAYER for SDL terminal interface
The HILITE_PLAYER option from config.d has been repurposed to highlight
the player character in the SDL terminal interface, because the SDL
terminal doesn't have a cursor.  Ergo, the init_player function will
check if the SDL terminal is being used and then decide whether or not
to highlight the player, but only if HILITE_PLAYER == true.

HILITE_PLAYER is now true by default.