Relicensing under Apache 2.0
Spelunk! is meant to be a base for larger projects, something that I and
other people can build off of and use to make greater things.  As such,
I've decided, with approval from my fellow contributor, to relicense it
as permissively as I feel I can allow.  Additionally, because I'm using
a standard license now, it will be easier for other users to determine
if their own licenses or third-party licenses are compatible.