Major revamps to color code
Because using the curses system for handling color and then
reinterpreting it later in the SDL interface was proving to be a hassle,
color has been genericized using a struct and that struct has been
implemented in the normal display code.  This not only made defining
colors a bit easier (though the back-end is a lot more explicit and
special-casey) it also fixed the problem of the SDL interface not
reversing colors properly.

Water is now blue.