Configuration options for SDL font and size
Added configuration options in config.d which determine the font and the
font size.  The Bold version of DejaVu Sans Mono has been added in order
to give the user the option of a bold font; this looks a bit better
because the colors are more solid, but it also gives the map a slightly
cluttered appearance.

The changelog has been updated.