Added a known bug to the changelog
A bug found by Elronnd in which the ncurses interface fails to draw
"seen" tiles outside of the player's field-of-view has been added to the
changelog.  Because we can't determine the source of the bug, but it
does not break the game (curses is no longer the default graphical
interface anyway), it has been documented pending further updates.