Messages and status bar now display in bold
Because neither the bold nor standard versions of the DejaVuSansMono
font are ideal on the small default display size, the two fonts are now
both used, each for the purpose that they work best for.  The result is
hopefully an ideal graphical experience for the user without necessarily
requiring a scale-up.