Added functionality for saving and loading levels
Functions have beed added which allow the saving and loading of levels
to a file, including revelant exception types.  This functionality is
currently not in use, as there is no mechanism for the player to save.

This took me a week to get right and now I'm not sure when the next time
I'll be able to do this is because my landlady is shutting down the
man-cave I'm using for an office.