Improved versioning system
The program now outputs the git commit number that it was compiled on
when asked for a version number, either by the --help prompt or by the
'v' key in-game.  A -v prompt has also been added to allow the user to
quickly look up the version number.  A flag has been added in config.d
which allows the user to disable this functionality if they are not
compiling from a git repository.

Additionally, a NAME flag has been added in global.d which will allow
users to easily change all references to SwashRL in case they want to
use the source code as the basis for their own game.

The readme file has been updated to reflect these changes, and to be a
little more helpful to people who are trying to compile their code.
Additionally, a table of contents has been added.