Compressed display functions
The display_map function has been simplified so that field-of-vision is
checked _after_ the rest of the function has already been completed.
This is an ugly fix, but it means that we don't have to duplicate as
much for FOV to work while still allowing for it to be switched on and
off.