Removed invent.d
Previously invent.d was a redundant source code file used because the
inven.d file was meant to store function declarations related to
inventory management but couldn't because the inventory needed to be
defined prior to the "player" struct.  Now, the functions defined in
invent.d have been distributed between inven.d and mov.d; the rationale
for this change is that "picking up" an item is a movement action and
therefore should be put in the file that defines movement functions,
whereas checking an item's equipment slot is a function designed to
process metadata in relation to the inventory, and thus it should be
declared alongside the inventory.

This was a longwinded way of saying a redundant source code file has
been removed.