Updated PDCurses warnings and bug reports
Due to feedback from our unofficial Linux maintainer it has come to my
attention that PDCurses is working on Windows, and with a little help
from unleashy on GitHub I have gotten the program to compile with
PDCurses, but any attempt to actually use the PDCurses interface causes
an error.

Pragma messages and the readme have been updated to reflect these new
developments, and links to the relevant Issue on GitHub have been
provided.