Removed PDCurses warnings
Following unleashy's comments on the GitHub issue wrt compiling with
PDCurses, it has become clear that PDCurses is now working for me, so
I've removed the warnings telling the user that it's broken and noone
can figure it out.

Now all I have to do is fix the curses interface.  Yay!