Fixed curses colors
Due to issues with the assignment of numeric IDs for curses color pairs,
the color pairs are now labeled with hard-coded values.

The original error that I was trying to fix seems to have been caused by
attempting to assign a color pair to the number 0, which for some reason
caused bright white to be assigned.  As I was trying to fix this more
color pairs became broken until I just decided to hard-code the values.

I did all of this while I was writing autodoc comments for iocurses.d,
so incidentally there are also now autodoc comments for iocurses.d.