Changed the player into a monst
In order to reduce the amount of work that goes into coding for monsters
as they get more and more similar to the player, the player character is
now defined by a monst struct rather than its own separate player
struct.  The monst struct has been updated to include items that were
previously unique to the player.