Updated struct & function names for consistency
All struct and function names have been updated to fit the standards for
the new style guide.  Previously the code contained a lot of structs and
functions which were not consistently named because a code style had not
been formally decided in the days before the source code was released.