Changed all integers to ints
All integers in function parameters and return types are now either int
or uint.  In fact most of the numeric values in the code are now either
int or uint, with the exception being coordinate values, which are
ubytes.  This has been done to reduce the amount of conversion necessary
to expand the codebase.