Check for an open grasp
A utility function has been added to item.d which allows to check if an
item is present, or if the given Item is in fact a placeholder.  This
can be used to bypass the cumbersome `Item.sym.ch == '\0'` check that we
do for floor tiles and inventory slots.

This utility function has been used to add a function to inven.d which
checks the player's hands to see if they have at least one open to grab
an item.

This `check_grasp` function has been used in `control_inventory` to
ensure that the player has a free hand before accessing their bag.

The bag has yet to be implemented.  A comedy of futility in three acts :-)