Further sketches of "bag" code
The "bag" appears to be mostly functional but suffers from some bugs wrt
actually outputting the result of a button press.  Delegating the "bag"
to its own function would probably be smart, because this whole process
is making the inventory management function easily the most complicated
individual function in the entire game, which I think says a lot about
the basic structure of the underlying code.

Good thing we're open-sourcing this stuff...