Moved equipment screen drawing into a new function
As part of my project to de-clutter and de-spaghettify the inventory
management code, the code to display the equipment screen has been
pushed into a separate function, SwashIO.display_equipment_screen.

Some idiosyncratic changes have been made to the code to make it more
clear what the intent of certain lines are.  The variables `line` and
`grabbed_line`, which previously had a default value of 255 to mean "the
player has not yet grabbed an item", now have a default value of -1 to
make this meaning more clear.  It will also enable users who want to
include more inventory slots to have an arbitrary maximum value for
these numbers.