Moved bag management to its own function
A new function SwashIO.manage_inventory manages the player's "bag," now
referred to internally as the "inventory."  This function also
instructs the equipment screen whether or not to redraw.

Not all of the functions related to "bag" management have been relegated
to this function; before manage_inventory is called, the
control_inventory function checks to make sure the player is actually
able to access their bag (e.g. they have an open grasp)

Some idiosyncratic changes have been made to the user interface to make
it more generally user-friendly and to reduce the amount of thinking
required by the user to deduce what the program wants them to do.