Fixed garbage output on the message line
A bug in which the SDL interface would display garbage output upon
starting the game has been fixed.  It was caused by an erroneous call to
`SwashIO.read_messages` during initialization.  Instead a call to
`SwashIO.clear_message_line` has been placed to ensure that the top line
will always be blank, but will still be filled in properly, at the start
of the program.