Fix greetings at start of game
The messages that "greet" the player at the start of the game have been
added back in.  This is done to re-implement the cheat code activation
notifications that previously existed but also to ensure that attempting
to read messages at the beginning of the game doesn't result in garbage
output.