Added a mapgen function for the program
A new command line argument `--mapgen` has been added which causes the
function to generate a map, display it, and then exit without starting a
proper game.  This can be used for testing map generation code or simply
generating sample maps.

A possible to-do for this addition is to add the ability to determine
what kind of map this function will generate, but since we're not that
far yet we'll let that slide.