Updated the readme with more up-to-date info
The "Supported Systems" section has been changed to specify that SwashRL
is not currently being maintained for any systems other than Linux and
why this is.

The "Known Bugs" section has been updated in order to reflect changes to
the code and to include a known bug which can occur as a result of the
recursive shadowcasting algorithm used in line-of-sight.