Mapgen now connects rooms randomly
Previously a mistake in the code caused the map generator to always
connect rooms with corridors in linear order, resulting in maps which
were very linear.  Now the shuffled room data is used properly, causing
the rooms to be connected in random order.

The way this is done in the code, with each room being connected to two
other rooms (with the possible exception of the very first and last
rooms) ensures that each room is accessible, and as an added bonus doing
this in random order results in more twisty passages and rooms being
more likely to have multiple doorways, which is certainly desireable in
a Roguelike.