Properly check equipment slots against equip types
Previously, the function `check_equip` would check a bitfield `equip` in
Item structs against an equipment slot to determine if the slot was a
valid match against the equipment type.

Due to a minor misunderstanding of how casting worked in D, the given
Item's equipment type was being cast as a boolean variable so that the
result could be returned by the function, when in fact the result of the
check should have been converted.  This has been fixed via the
application of parethesis.