Greet the player during holidays
On certain significant dates, the game will now give the player an
appropriate greeting on startup, in addition to the usual "Welcome back
to SwashRL!" message.

The greetings are as follows, sorted from latest to earliest in the
year:

December 31-Jan 1: New Year's Eve/Day, an international celebration of
the coming of the new year on the Gregorian calendar.  Both dates are
marked on most calendars because the new year is traditionally held to
begin at midnight in-between the two dates.
  "Happy New Year!"

December 24-25: Christmas Eve/Day, a time of sharing, family reunions,
and merriment.  The holiday itself is a largely secularized version of
a Christianized form of Yule and other pagan festivals traditionally
held on the darkest days of the year in Europe.
  "Merry Christmas!"

October 31: Halloween, a secularized version of an ancient
Pagan/Christian harvest festival, now primarily celebrated in the United
States as an excuse to dress up and eat candy.
  "Happy Halloween!"

April 8: Hanami, the Japanese Cherry Blossom Festival, which typically
is held throughout the month of April but for this game's purposes is
set during Hanamatsuri, a festival which additionally celebrates the
birth of the Buddha.
  "Happy Hanami!"