Add a "drop" command to the keymap function
The function which defines keymaps now contains a key which represents
the "drop" command, 'd'.

As has been done in the past, this command has been added in the same
order as it appears in the moves.d file, and all commands in front of it
in the function have been shifted over accordingly.