Add the ability for a "drop" action to occur
A function `put_down` has been added to `mov.d` which will cause a
monster to put down the item at the given index, or if the monster is
the player and the index is negative allow the player to choose an item
to drop.

This function has not yet been written and instead just displays a
message.  The important thing about this commit is that the function
can now be called by the umove routine.