Sketch the put_down function
The put_down function _should_ now remove an item from the inventory
and return it *if* `index` is non-negative and is a valid index for
the inventory items array.

It still does not allow any player feedback, so functionally the player
should not notice any difference between this and the previous commit.

The next steps are to write a function which will displaye the
equipment screen and allow the player to select an item, which
shouldn't be too much of a hassle, right?  ...right?