Restructure put_down & correct casting error
The previous commit introduced a casting error which was causing the
put_down function to believe that the default index value of -1 was
greater than the length of the inventory array.  This commit corrects
this error by using the hard-coded maximum possible value of the
inventory items array (40) and by restructuring the program's flow so
that it checks for negative values first.