Change put_down into m_drop_item
It's become obvious that the put_down function as currently written
will not work, because in order to drop an item from the equipment
screen the `IO` object has to be defined, but it's not defined until
`main` runs, so we need to defer that behavior to a function in `main`
and may need to declare a player-specific "drop" function in the
SwashIO template itself.

As such, the existing `put_down` function has been changed to
`m_drop_item`, a function specifically for having a monster drop an
item from its inventory.