Add an "invert" function to Color_Pair
To make it easier to invert colors in the display functions, the
`Color_Pair` class has been given a function `invert`, which returns a
new `Color_Pair` with reversed colors.

Every `Color_Pair` now also has a boolean `inverted`, which much like
`bold` is used to signal that a color pair should be inverted in the
curses interface.