Have iocurses.d use the new color system
All functions in iocurses.d which reference colors in any way now make
use of either a `Color_Pair` or the `Colors` enum.

The `get_color` function has been changed to take in a `Color_Pair` and
convert it into the result of a `COLOR_PAIR` call.

An unnecessary operator has been taken out which checks to see if
`COLOR` is true.