Split Color_Pair.brighten into two functions
Since we can't use negative values to tell the "brighten" function we
want it to brighten the SDL colors dynamically, I've used overloading to
split it into two functions, one of which requires RGB specifications
and one of which doesn't.  I've gone ahead and added a check for
overloading as well.