Fully rewrite global color declarations
Because I didn't understand how enums and static variables work, I had
to abandon the `Colors` enum I worked so hard on and instead have all of
the standard color pairs be declared statically and initialized with a
function, which is now called in main.  All references to the `Colors`
enum have now been replaced with `CLR_*` values like in the old code
base.

I'm not salty about this at all, stop staring at me.