Replace `Terrain` enum w/ static variables
Due to the previous commit, in which we defined color pairs using static
variables, we now must define standard terrain elements in the same way
because we can not initialize them properly at compile time.

All references to terrain elements in the source code have been replaced
with TERRAIN_* calls.

I'm still not salty about this.