curses_color_pair now exists regardless of curses
The `curses_color_pair` & `last_color_pair` variables in the
`Color_Pair` class, and the functions used to access them, now exist
regardless of whether curses is included or not.

The reason for this is that these curses color pair values are useful
for identifying the color value later using a numerical identifier, and
it would be foolish to squander that opportunity.