Add screenshots to the readme
Screenshots of a game in-progress have been added to the readme file.

The first is a screenshot of a game having just loaded in a saved file,
which appears at the very top of the readme.

The second is the same screenshot as before but using the OpenDyslexic
font, which appears in the "Dyslexia-Friendly Font" section.