Create initial sketch of fonts.d
A new file, fonts.d, has been added, which defines a struct used to
store where a font is accessed and the appropriate tile height and width
for that font.

All of the code inside the new file is wrapped in a `version( sdl )`
specification so that it is not read by versions which do not make use
of it.