swap curses definitions for black & dark gray
A misunderstanding of how curses defines "black" resulted in dark gray
being used instead of black and vice-versa.  The color definitions have
been swapped and CURSES_DARK has been renamed to CURSES_BLACK to avoid
future confusion.

This commit fixes the error in the test map which caused the goobling to
display as a pure black rather than dark gray.