cause player to appear in unique-ish color
The player character is now displayed in a unique orange color in the
SDL interface to make them stand out.

Because many terminals' "bright red" color is in fact bright red, rather
than orange, the player is instead displayed in white in the curses
interface, as this would be less confusing and present a friendlier
appearance.