use better hiliting on equipment screen
In early versions of the game, a grabbed item in the equipment screen
would have been inverted, rather than having that "GRABBED" line.  Now
it's much easier for us to do that, so do that instead.

Additionally, empty slots will use a distinctive but undistracting
highlight to indicate that they are available.