use enums to choose directions
Rather than using the Move enum for choices of directions to move game
objects, a more appropriate Direction enum has been written for this
purpose, with a function that we can use to get an appropriate
Direction from a Move.

As a consequence of this change, certain player movement functions which
were previously allocated to the `umove` function have been moved to the
mainloop.