remove references to deleted symdata function
The `symdata` function was removed in the previous commit for having
been deprecated for a substantial period of time.  This commit
replaces all references to it with the `Symbol` constructor as
suggested by the comment block that was there previously.

Because tsyms.d is so short, this commit goes ahead and conforms it to
the style guide.