rename Symbol member `ch` per style guide
This is another file where I forgot to check for compliance with the
three-character variable name minimum, and `Item.ch` was kind of a worst
offender.  It has been renamed to `ascii` in reference to the fact that
the variable is intended to specifically represent the symbol's
representation on ASCII text displays.