correct historical copyright years in source files
The generic 2015-2020 copyright year used in most of the source files is
based on the 2015-2020 date given in LICENSE.txt to refer to SwashRL as
a whole.  However, this is not accurate for every source code file; the
majority of them are descended from files which were introduced in 2017
and some of them were first written just earlier this year.

From here on out each source code file will be listed as being
copyrighted from their date of origin just for bookkeeping purposes.

Files which are descended from the original C implementation of SwashRL
mostly use the same names that they had at the time of the switch-over,
making the linearity of the source code history fairly easy to verify.