cause umove to exit after detecting a collision
The `monster` and `terrain` variables were not being used to stop
collision detection or movement in the `umove` function, which caused a
bug where the player would drown immediately after attacking a monster
that was sitting on a water tile, even though the player was not
supposed to move onto that tile.

To fix these issues, the `monster` and `terrain` variables have been
removed and `umove` now returns on the first detected collision whether
it results in a turn being expended or not.