add functions for properly displaying item names.
The `item_name` function will return an appropriate name for a given
item, and will pluralize that name if the item is a stack of multiple
items.  If `count` is true, it will include either a count of the items
in question or an "a" to indicate how many items there are.

The `count_items` function functions like `item_name` with `count` set
to true.