Revert "update version number to 0.∞"
This reverts commit 01c4d5315baa703300ac4706c72f8cb191e466a5.

I decided, in hindsight, that an infinite alpha development model would
be dumb, even for a project as slow-moving as this, and that I have a
lot more fun treating SwashRL like a continuously updating project with
an actual version number system.  As such, I'm going back to the old
system.