generalize writing the status bar.
A new function `write_status_bar` has been written in Swash_IO which
writes the text for the status bar.  `refresh_status_bar` is still used
to call this function.

This change was made because any change to the format of the status bar,
such as the change made by the addition of Strength & Endurance stats,
would need to be written twice otherwise, because the curses and SDL
terminal interfaces need to do different setups before writing the
status bar.

Now, instead, the status bar is formatted in `write_status_bar` and
`refresh_status_bar` does display-unique setup.