replace armor types with equipment slots.
The `Slot` enum is being set up as the replacement for both the `Armor`
enum and the `INVENT_*` constants to refactor the equipment screen and
inventory code.

In future versions of the code, the `Slot` values will be used both to
label equipment slots and to determine where armor or other items can be
equipped.