separate monster equipment & inventory.
The full list of changes that have been made by this commit is truly too
much for me to concisely explain in this commit message.  The long and
short of it is that a monster's inventory is now separated into three
separate arrays:
 - inventory - a dynamic array of `Item` structs representing the open
   inventory.
 - wallet - an array of five `Item` structs used to store money in five
   different denominations.
 - equipment - an associative array of `Item` structs labelled with
   `Slot` values representing items either equipped or held in hands.

The number of equipment slots has been reduced.  See the changelog for a
comprehensive list.  The test items in the test map have been altered to
reflect this.

The player no longer gets a festive hat if they are already wearing a
hat at the start of the game.

There is no longer an `Inven` struct.

This commit took me two days to compose and I'm very tired, so I
apologize if this is not a comprehensive explanation of changes but
frankly I'm running out of cares.