implement interpretation of bd-font data.
bd-font is implemented differently than the TTF files that SwashRL
normally uses; rather than being a TTF file, it is defined as an array
of bits which are interpreted in the `SDL_Terminal_IO` class to create a
tile set.  This method allows compiled binaries to have a default font
which can be used in the event that an invalid font file is specified by
the user.